maybe solved codeclimate issues
